% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinInfo.R
\name{addDataToProtein}
\alias{addDataToProtein}
\title{addDataToProtein}
\usage{
addDataToProtein(
  proteinDF,
  dataframe,
  dataColumn,
  dataName,
  positionColumn = "position",
  NAValue = 0L,
  strictPosition = TRUE
)
}
\arguments{
\item{proteinDF}{the protein data.frame to add the info to. Main requirements are that there is a column named
position, that there is no column named 'newData____' (internal use) and no column with the same name as defined
by the 'dataName' argument}

\item{dataframe}{data.frame with the data to be added. Should have two columns: one with the position info and
another with the data relating to these positions}

\item{dataColumn}{name of the column in dataframe with the data to be added}

\item{dataName}{name of the new data column in the protein data data.frame}

\item{positionColumn}{name of the position column in dataframe which has the position info (default is 'position')}

\item{NAValue}{value for the position in the protein data data.frame which are missing in the (new) dataframe data}

\item{strictPosition}{logical vector. Default is TRUE: if the positions in the (new) dataframe are outside the range
of the protein data, then an error will be generated. If FALSE, then only a warning is generated}
}
\value{
data.frame
}
\description{
Takes data.frame with position data and adds it as an extra column to a protein data data.frame
Positions not covered in the added data.frame are set to the argument NAValue (default is integer 0)
}
\examples{
bsaProtein <- createProteinData(sequence = standardProtein("bsa"), nterm = 1,
 start = 25)
toAdd <- data.frame(position = c(25:30, 100:110), data = c(rep(1,10),
 rep(3, 3),rep(2,4)))
bsaProtein <- addDataToProtein(proteinDF = bsaProtein, dataframe = toAdd,
 dataColumn = "data", dataName = "found")
bsaProtein$position[bsaProtein$sequence == "K" | bsaProtein$sequence == "R"]
trypticSites <- data.frame(position = bsaProtein$position[
 bsaProtein$sequence == "K" | bsaProtein$sequence == "R"],
 data = "tryptic")
bsaProtein <- addDataToProtein(proteinDF = bsaProtein, dataframe = trypticSites,
 dataColumn = "data", dataName = "trypticSite", NAValue = "")
bsaProtein[95:110,]
}
