% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{boxData}
\alias{boxData}
\title{boxData}
\usage{
boxData(
  data,
  ncols = NA,
  adjustRows = TRUE,
  emptySequence = "",
  replaceNA = NA
)
}
\arguments{
\item{data}{data.frame, should not already have a 'row' and/or 'col' column}

\item{ncols}{number of required columns for the matrix}

\item{adjustRows}{logical vector, default is TRUE. Defines whether the number of rows for the data.frame is allowed to be adjusted.
Setting this to FALSE may lead to 'stop'-error if data doesn't fit the requested matrix.}

\item{emptySequence}{if data.frame is extended, this character vector is used to sequence in the new rows. The position column is simply
filled with increasing integers.}

\item{replaceNA}{default is NA, if defined as something else, then ALL NA values in the data.frame will be replace with this argument.
This is an admittedly crude way to replace NA values. If more finesse is needed, leave at default and handle NA values in followup functions
such as 'replace_na' from the tidyr package}
}
\value{
a data.frame with extra columns (row and col) and possible extra rows
}
\description{
Takes a data.frame and adds a 'row' and 'col' column which together define which cell in the rows*columns-matrix a data.frame-row maps to
}
\note{
If the data.frame extended: the position column and the sequence are filled with increasing integers (starting from the max position)
and the 'emptySequence' argument respectively. Other columns in the extended using NA
}
\examples{
boxData(createProteinData(standardProtein(), start = 2, nterm = 1), ncols = 50) |> head(20)
boxData(createProteinData(standardProtein(), start = 2, nterm = 1), ncols = 50) |> tail(20)
boxData(createProteinData(standardProtein(), start = 2, nterm = 1), ncols = 50)[40:60,]
}
