% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{boxDimensions}
\alias{boxDimensions}
\title{boxDimensions}
\usage{
boxDimensions(
  vectorLength = 100,
  ratio = NA,
  increaseColumns = FALSE,
  ncol = NA,
  nrow = NA,
  noWarning = FALSE
)
}
\arguments{
\item{vectorLength}{minimum number of cells for the resulting columns*rows matrix}

\item{ratio}{the intended ratio the resulting columns*rows matrix. Increase ratio to get more columns, decrease to get more rows.}

\item{increaseColumns}{if TRUE, then the number of columns will be increased to reach the number of matrix cells required.
If FALSE then the number of rows will be increased to do this. Default is FALSE}

\item{ncol}{defines the number of columns to use. If not also defined, the number of rows is adapted to fit all of the required number of cells for the matrix}

\item{nrow}{defines the number of rows to use.  If not also defined, the number of columns is adapted to fit all of the required number of cells for the matrix}

\item{noWarning}{logical vector, if product of the number of rows and the number of columns is too small for the number of matrix cells (vector length) required
then a warning will displayed (default noWarning is FALSE), unless noWarning is TRUE}
}
\value{
a named numeric vector (names: ncol, nrow)
}
\description{
Calculates/defines the dimensions of the 'matrix' to use to display data of a certain length (e.g. a protein sequence)
}
\note{
Order of importance for the ratio, ncol and nrow arguments: if the ratio is NA (default), then the ncol argument
is used, if possible with the nrow argument. If ncol is NA, then the nrow argument is used.
}
\examples{
boxDimensions(385, 5)
boxDimensions(385, nrow = 9)
boxDimensions(400, 10)
boxDimensions(400, ncol = 60)
}
