% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{containsModification}
\alias{containsModification}
\title{containsModification}
\usage{
containsModification(
  modifications,
  whichModification,
  strict = TRUE,
  splitCharacter = ",",
  trim = TRUE
)
}
\arguments{
\item{modifications}{character vector of format \if{html}{\out{<modification>}} \if{html}{\out{<splitCahracter>}} \if{html}{\out{<modification>}} \if{html}{\out{<splitCahracter>}} etc etc}

\item{whichModification}{the modification to search for (character vector)}

\item{strict}{logical vector, default is TRUE: the search for the modification is strict. If FALSE then the search goes via the 'grepl' function
and regular expressions can be used}

\item{splitCharacter}{character vector used to split the modifications argument}

\item{trim}{logical vector. Default is TRUE: the modifications will be trimmed (spaces on both sides removed). When FALSE, this is not done}
}
\value{
logical vector
}
\description{
Helper function to check whether a character vector of format modification,splitCahracter,modification,splitCahracter, (etc etc) contains
a certain modification
}
\examples{
containsModification(c('Oxidation,Phospho','Carbamidomethyl,Phospho'),
 whichModification = "Phospho")
containsModification(c('Oxidation , Phospho','Carbamidomethyl, Phospho'),
 whichModification = "Phospho")
containsModification(c('Oxidation , Phospho','Carbamidomethyl, Phospho'),
 whichModification = "Phospho", trim = TRUE)
containsModification(c('Oxidation , Phospho','Carbamidomethyl, Phospho'),
 whichModification = "Phospho", strict = FALSE)
}
