% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinInfo.R
\name{createProteinData}
\alias{createProteinData}
\title{createProteinData}
\usage{
createProteinData(
  sequence,
  start = 1,
  end = nchar(sequence),
  nterm = NA,
  cterm = NA,
  emptySequence = " "
)
}
\arguments{
\item{sequence}{character vector, protein sequence in the form of the usual letter codes}

\item{start}{first position in the sequence to be used}

\item{end}{last position in the sequence to be used}

\item{nterm}{start position 'adjustment'. Adds a number of 'empty' rows to the head of the resulting data.frame to have it start at position nterm}

\item{cterm}{end position 'adjustment'. Adds a number of 'empty' rows to the tail of the resulting data.frame to have it ends at position cterm}

\item{emptySequence}{default " ", character vector for the sequence column to use when adding rows. Note: other columns will also be
extended and usually add NA's as a value}
}
\value{
data.frame with two columns: position (integer vector) and sequence (amino acid letters: character vector)
}
\description{
Create a protein data.frame with basic information: position & sequence. In the resulting data.frame, every row
has an amino acid (letter, sequence column) and the position of that amino acid in the protein (position column)
}
\note{
nterm stands for n-terminus, cterm stands for c-terminus
}
\examples{
createProteinData(sequence = standardProtein("bsa"), start = 25) |> head(25)
createProteinData(sequence = standardProtein("bsa"), nterm = 1, start = 25) |> head(25)
createProteinData(sequence = standardProtein("bsa"), start = 25) |> tail(10)
createProteinData(sequence = standardProtein("bsa"), start = 25, nterm = 1, cterm = 610) |>
 tail(10)
}
