% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{dfToModPositions}
\alias{dfToModPositions}
\title{dfToModPositions}
\usage{
dfToModPositions(df, splitChar = ",", returnEmptyRow = TRUE)
}
\arguments{
\item{df}{a single row data.frame having the 'collapsed' columns (modifications, positions, letters and numbers)
Each modification(-type) is separated by the argument 'splitChar'}

\item{splitChar}{character vector specifying the separation 'character' that is placed between the modification definitions}

\item{returnEmptyRow}{logical vector, default is TRUE. TRUE: if no modifications are present in the 'df' argument, a data.frame will
be returned without rows. FALSE: in that case a data.frame with a single row, with in every column an empty character vector, will be returned.}
}
\value{
a data.frame with columns: modifications, positions, letters and numbers
}
\description{
Takes a single modification table row with one or more modification definitions (as produced by \code{\link{modPositionsToDF}})
and makes it into a single-modification-per-row data.frame. A row with 'x' modifications becomes 'x' number of rows. Essentially
the opposite of \code{\link{modPositionsToDF}}.
}
\examples{
modPositions(OVATable("peptide")$Modifications[32]) |> modPositionsToDF() |> dfToModPositions()
modPositions(OVATable("peptide")$Modifications[2]) |> modPositionsToDF() |> dfToModPositions()
modPositions(OVATable("peptide")$Modifications[1])
modPositions(OVATable("peptide")$Modifications[1]) |> modPositionsToDF()
modPositions(OVATable("peptide")$Modifications[1]) |> modPositionsToDF() |> dfToModPositions()
}
