% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinInfo.R
\name{endPosition}
\alias{endPosition}
\title{endPosition}
\usage{
endPosition(string, splitCharacter = "-")
}
\arguments{
\item{string}{character vector with format 'xxx-yyy', where xxx is the start
position and yyy is the end position}

\item{splitCharacter}{default is "-", which is used by Proteome Discoverer}
}
\value{
numeric vector
}
\description{
Gives the end position of a character vector with format 'xxx-yyy'.
Convenience function
}
\examples{
endPosition("901-902")
endPosition("1-102")
}
