% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{getPositionLetters}
\alias{getPositionLetters}
\title{getPositionLetters}
\usage{
getPositionLetters(position)
}
\arguments{
\item{position}{character vector with format \if{html}{\out{<letter>}}\if{html}{\out{<number>}};\if{html}{\out{<letter>}}\if{html}{\out{<number>}};... etc etc}
}
\value{
character vector with format \if{html}{\out{<letter>}};\if{html}{\out{<letter>}};... etc etc
}
\description{
helper function which extracts the letter codes (for amino acids) from the position info part of a modification
}
\note{
internal function
}
