% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{getPositionNumbers}
\alias{getPositionNumbers}
\title{getPositionNumbers}
\usage{
getPositionNumbers(position)
}
\arguments{
\item{position}{character vector with format \if{html}{\out{<letter>}}\if{html}{\out{<number>}};\if{html}{\out{<letter>}}\if{html}{\out{<number>}};... etc etc}
}
\value{
character vector with format \if{html}{\out{<letter>}};\if{html}{\out{<letter>}};... etc etc
}
\description{
helper function which extracts the positions from the position info part of a modification
}
\note{
internal function

position info which does not have a letter will return a zero as position
}
