% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinInfo.R
\name{getPositions}
\alias{getPositions}
\title{getPositions}
\usage{
getPositions(string, Accession, exact = TRUE, splitCharacter = ";")
}
\arguments{
\item{string}{character vector in the format: 'Accession' 'position'; 'Accession' 'position'; etc}

\item{Accession}{character vector that specifies from which protein Accession to get the position}

\item{exact}{logical vector which defines how the Accession argument is compared to the string argument.
Default is TRUE which means that the Accession needs to be exactly the same letters/numbers.
If FALSE then a 'grepl' statement is used with the Accession as a pattern}

\item{splitCharacter}{character vector which specifies the beginning/end of 'Accession' 'position' pair.
Default is ";" which is used by Proteome Discoverer}
}
\value{
a character vector containing the position of the specified Accession
}
\description{
Gets the part of the string containing position information. Works with
data organized in the 'PositionsinProteins' and/or 'PositionsinMasterProteins'
column of peptide tables as defined in Proteome Discoverer
}
\note{
this function expects Accessions to be unique. If this is not true or the parameter exact is set to FALSE
it may return character vectors with length > 1
}
\examples{
getPositions("P00000 [352-360]; P01012 [351-359]", Accession = "P01012")
getPositions("P00000 [352-360]; P01012 [351-359]", Accession = "P00000")
getPositions("P00000 [352-360]; P01012 [351-359]", Accession = "P0", exact = FALSE)
}
