% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{graphsAdjust}
\alias{graphsAdjust}
\title{graphsAdjust}
\usage{
graphsAdjust(
  graphs,
  vertical = FALSE,
  xDiscrete = FALSE,
  yDiscrete = FALSE,
  xReverse = FALSE,
  yReverse = FALSE,
  xDefault = FALSE,
  yDefault = FALSE,
  xLimits = c(0, NA),
  yLimits = c(0, NA),
  xExpand = ggplot2::expansion(mult = 0, add = 0),
  yExpand = ggplot2::expansion(mult = 0, add = 0),
  xLabelFormat = ggplot2::waiver(),
  yLabelFormat = ggplot2::waiver(),
  xOob = scales::oob_squish_infinite,
  yOob = scales::oob_squish_infinite,
  xLog = FALSE,
  yLog = FALSE,
  xIsDate = FALSE,
  yIsDate = FALSE,
  titles = NA,
  xLabel = NA,
  yLabel = NA,
  setTheme = theme_minimal_adapted(),
  plot.margins.default = TRUE,
  plot.margins = c(5, 5, 5, 5),
  plot.margins.units = "points"
)
}
\arguments{
\item{graphs}{list of ggplot-objects to which the adjustments have to be made
Note: MUST be a list}

\item{vertical}{if TRUE, flips x- and y-axis}

\item{xDiscrete}{specifies whether the x-axis should be discrete}

\item{yDiscrete}{specifies whether the y-axis should be discrete}

\item{xReverse}{specifies whether to reverse the x-axis}

\item{yReverse}{specifies whether to reverse the y-axis}

\item{xDefault}{if TRUE, then xExpand, xLimits and xOob are ignored
(essentially autoscaling the x-axis)}

\item{yDefault}{same as xDefault, but for y-axis}

\item{xLimits}{range of the x-axis, normally xLimits = c(minimum, maximum)
if minimum and/or maximum is NA, then they are autoscaled, if xLimits = NA
then the range is 0, putting all datapoints in one line (x-axis-wise)}

\item{yLimits}{range of the y-axis (see xLimits)}

\item{xExpand}{allows for padding around data (x-axis),
see ?ggplot2::expansion for proper explanation}

\item{yExpand}{allows for padding around data (y-axis),
see ?ggplot2::expansion for proper explanation}

\item{xLabelFormat}{defines the numeric format to be used for the x-axis
labels (see fromatDigits() & formatScientificDigits() for examples)}

\item{yLabelFormat}{defines the numeric format to be used for the y-axis
labels (see fromatDigits() & formatScientificDigits() for examples)}

\item{xOob}{defines what to do with data that's out of range of the data,
see ?scales::oob for proper explanation. Note: only deals with x-axis}

\item{yOob}{defines what to do with data that's out of range of the data,
see ?scales::oob for proper explanation. Note: only deals with y-axis}

\item{xLog}{if TRUE then automatic transformation of the x-axis to logarihmic
scale}

\item{yLog}{if TRUE then automatic transformation of the y-axis to logarihmic
scale}

\item{xIsDate}{if TRUE then all settings regarding the x-axis are ignored,
except xLimits which should be dates (defaults are not dates). Alternatively
xDefault can be set to TRUE for autoscaling}

\item{yIsDate}{if TRUE then all settings regarding the y-axis are ignored
except yLimits which should be dates (defaults are not dates). Alternatively
xDefault can be set to TRUE for autoscaling}

\item{titles}{sets title of graph}

\item{xLabel}{sets x-axis title}

\item{yLabel}{set y-axos title}

\item{setTheme}{if NA, then no theme is applied, otherwise uses the defined
theme (can also be ggplot2 included themes, such as theme_bw())}

\item{plot.margins.default}{if TRUE, ignore plot.margins and other parameters}

\item{plot.margins}{defines margins (from the border) of the plot
c(top, right, bottom, left)}

\item{plot.margins.units}{default = "points", other possibilities:
?grid::unit , examples "cm", "points", "inches", "npc" (viewport) etc etc}
}
\value{
a list of ggplot objects
}
\description{
Make adjustments to a graph eg zoom, titles etc etc
}
\note{
also part of the personal package 'BBPersonalR'
}
