% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ifelseProper}
\alias{ifelseProper}
\title{ifelse replacement for properly returning all datatypes.}
\usage{
ifelseProper(logicValue = NULL, ifTrue = NULL, ifFalse = NULL)
}
\arguments{
\item{logicValue}{variable or expression resulting in TRUE or FALSE,
if missing or not logical then the function will return NULL.}

\item{ifTrue}{variable or expression to be returned when logicValue == TRUE}

\item{ifFalse}{variable or expression to be returned when logicValue == FALSE}
}
\value{
depending on logicValue, ifTrue ir ifFalse.
}
\description{
ifelse replacement for properly returning all datatypes.
}
\note{
not vectorized

internal function
}
