% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is.Class}
\alias{is.Class}
\title{is.Class}
\usage{
is.Class(object, whichClass)
}
\arguments{
\item{object}{a data object of some class}

\item{whichClass}{character string: class name to be tested}
}
\value{
TRUE or FALSE
}
\description{
helper function to determine if object == whichClass or a descendant
of whichClass.
}
\note{
internal function
}
