% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein3d.R
\name{log10add}
\alias{log10add}
\title{log10add}
\usage{
log10add(valueToAdd)
}
\arguments{
\item{valueToAdd}{numeric value to add to each value before taking the log10 in the resulting
function}
}
\value{
function
}
\description{
function factoru that generates a function that takes a numeric value, adds a
specified value and then gives the logarithmic value with base 10. This is similair tp
\code{\link{log1p}}. It prevents problems with zero. It does add a small value to
the original value, but as this is here meant for log10 transformation of values over a
large range (like Abundances of peptides in a digest), it should not be a problem.
}
\examples{
log10(0:10)
log10add(1)(0:10)
log10add(0)(0:10)
log10(c(0,1E5, 1E7, 1E8))
log10add(1)(c(0,1E5, 1E7, 1E8))
log10add(1E3)(c(0,1E5, 1E7, 1E8))
}
