% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein3d.R
\name{log10p}
\alias{log10p}
\title{log10p}
\usage{
log10p(value)
}
\arguments{
\item{value}{numeric value}
}
\value{
numeric value
}
\description{
takes a numeric value, adds 1 and then gives the logarithmic value with base 10.
This is essentially the log10 version of \code{\link{log1p}}. This prevents problems with
zero. It does add a small value to the original value, but as this is here meant for log10
transformation of values over a large range (like Abundances of peptides in a digest),
so it should not be a problem.
}
\examples{
log10(0:10)
log10p(0:10)
log10(c(0,1E5, 1E7, 1E8))
log10p(c(0,1E5, 1E7, 1E8))
}
