% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{modPositionsToDF}
\alias{modPositionsToDF}
\title{modPositionsToDF}
\usage{
modPositionsToDF(mods, collapseChar = ",", returnEmptyRow = TRUE)
}
\arguments{
\item{mods}{data.frame coming from \code{\link{modPositions}}}

\item{collapseChar}{character vector (default ',') which is used to collapse the columns of the rows (needed to undo
the 'collapse' in \code{\link{dfToModPositions}})}

\item{returnEmptyRow}{logical vector, default is TRUE. TRUE: if no modification is present in the 'mods' argument, a data.frame will
be returned without rows. FALSE: in that case a data.frame with a single row, with in every column an empty character vector, will be returned.}
}
\value{
a data.frame with columns: modifications, positions, letters and numbers
}
\description{
helper function that 'collapses' two or more rows in a modification data.frane (coming from \code{\link{modPositions}})
into a single data.frame row
}
\examples{
OVATable("peptide")$Modifications[32]
modPositions(OVATable("peptide")$Modifications[32])
modPositions(OVATable("peptide")$Modifications[32]) |> modPositionsToDF()
OVATable("peptide")$Modifications[2]
modPositions(OVATable("peptide")$Modifications[2])
modPositions(OVATable("peptide")$Modifications[2]) |> modPositionsToDF()
}
