% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{OVATable}
\alias{OVATable}
\title{OVATable}
\usage{
OVATable(whichone = "protein")
}
\arguments{
\item{whichone}{character vector which specifies what to get: 'protein' for a protein table, 'peptide' for a peptide table}
}
\value{
data.frame or NA
}
\description{
Helper function for examples & testing. Generates a data.frame, of protein or peptide information (both Ovalbumin), which is similar to
what comes out of Proteome Discoverer
}
\examples{
OVATable("protein")
OVATable("Peptide")
OVATable("Info")
}
