% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinInfo.R
\name{prepare}
\alias{prepare}
\title{prepare}
\usage{
prepare(
  data,
  start = min(data$position),
  end = max(data$position),
  emptySequence = " "
)
}
\arguments{
\item{data}{data.frame with two columns: position (integer vector) and sequence (amino acid letters: character vector)}

\item{start}{integer, should be less than the start position (in first row of data): adds positions with 'emptySequence' to the head of data}

\item{end}{integer, should be maximum positione: adds positions with 'emptySequence' to the tail of data}

\item{emptySequence}{default " ", character vector for the sequence column to use when adding rows. Note: other columns will also be
extended and usually add NA's as a value}
}
\value{
data.frame with two columns: position (integer vector) and sequence (amino acid letters: character vector)
}
\description{
Helper function to add (empty) rows to proteinData,
}
\examples{
createProteinData(sequence = standardProtein("bsa"), start = 25) |> prepare(start = 1) |> head()
createProteinData(sequence = standardProtein("bsa"), start = 25) |> prepare(end = 610) |> tail()
}
