% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein3d.R
\name{reduceColors}
\alias{reduceColors}
\title{reduceColors}
\usage{
reduceColors(proteinDF, positionColumn = "position", colorColumn)
}
\arguments{
\item{proteinDF}{protein data data.frame}

\item{positionColumn}{name of the column containing the position information}

\item{colorColumn}{name of the column containing the color information}
}
\value{
data.frame with three columns: start, end (both positions) and color
}
\description{
takes protein data with color information and creates a smaller data.frame contaning
three columns: start, end (both positions) and color. This can be used for mapping the colors
onto the 3D object. It is used inside the
}
\note{
internal function
}
