% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rowsAndcolumns}
\alias{rowsAndcolumns}
\title{rowAndColumns}
\usage{
rowsAndcolumns(
  vectorLength,
  ratio,
  increaseColumns = FALSE,
  minimumRatio = 0.01,
  maximumRatio = 100
)
}
\arguments{
\item{vectorLength}{minimum number of cells for the resulting columns*rows matrix}

\item{ratio}{the intended ratio the resulting columns*rows matrix. Increase ratio to get more columns, decrease to get more rows.
Note: ratio = number of columns divided by the number of rows.}

\item{increaseColumns}{if TRUE, then the number of columns will be increased to reach the number of matrix cells required.
If FALSE then the number of rows will be increased to do this. Default is FALSE}

\item{minimumRatio}{due to the iterative nature of part of the function, this puts a minimum limit on the possible ratios
to prevent the function from 'getting stuck'. Default is 0.01}

\item{maximumRatio}{due to the iterative nature of part of the function, this puts a maximum limit on the possible ratios
to prevent the function from 'getting stuck'. Default is 100}
}
\value{
a named numeric vector (names: columns, rows)
}
\description{
Calculates number of columns and rows of a matrix from number of cells and the ratio number of columns/ number of rows
The number of cells is a defined as a minimum number of cells. Either the number of rows or the number of columns is
increased to accommodate this. The column/row ratio and number of cells for the 'solution' might therefore not be exact.
The reason for this is that the intended target for this function is to come up with a cell matrix big enough for a
data of a certain size (vectorLength)
}
\examples{
rowsAndcolumns(385, 5)
rowsAndcolumns(385, 0.2)
rowsAndcolumns(400, 10)
rowsAndcolumns(400, 0.1)
}
