% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{standardProtein}
\alias{standardProtein}
\title{standardProtein}
\usage{
standardProtein(short = "OVA")
}
\arguments{
\item{short}{shortened name of the protein to retrieve. Only 'bsa' and 'ova' (or their upper case equivalents) are
valid inputs. Other input will return NA}
}
\value{
an amino acid sequence in the form of a character vector or NA
}
\description{
Helper function for examples & testing. Generate a character vector of the amino acid sequence
of one of the (2) standard proteins.
}
\note{
The data is coming from \href{https://www.uniprot.org/uniprotkb/}{UniProt}

OVA is short for \href{https://www.uniprot.org/uniprotkb/P01012/entry}{P01012 · OVAL_CHICK} Ovalbumin (Gallus gallus)

BSA is short for \href{https://www.uniprot.org/uniprotkb/P01012/entry}{P02769 · ALBU_BOVIN} Bovine Serum Albumin (Bos taurus)
}
\examples{
standardProtein()
standardProtein("bsa")
standardProtein("Insulin")
}
