% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinInfo.R
\name{startPosition}
\alias{startPosition}
\title{startPosition}
\usage{
startPosition(string, splitCharacter = "-")
}
\arguments{
\item{string}{character vector with format 'xxx-yyy', where xxx is the start
position and yyy is the end position}

\item{splitCharacter}{default is "-", which is used by Proteome Discoverer}
}
\value{
numeric vector
}
\description{
Gives the start position of a character vector with format 'xxx-yyy'.
Convenience function
}
\examples{
startPosition("901-902")
startPosition("1-102")
}
