% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{strMultiReplaceAll}
\alias{strMultiReplaceAll}
\title{strMultiReplaceAll}
\usage{
strMultiReplaceAll(strings, patterns, replacements = "")
}
\arguments{
\item{strings}{character vector (can be longer than 1) in which all the patterns (arguments) are to be replaced}

\item{patterns}{character vector of patterns to be replaced}

\item{replacements}{character vector of replacements for the patterns in the strings argument (if found). Should be
either length = 1 or same length as the replacements. Default is "" (empty character vector), which means that
all patterns in strings will be removed}
}
\value{
character vector
}
\description{
helper function that replaces multiple patterns in a character vector, essentially an extension
of stringr::str_replace_all
}
\note{
internal function
}
