% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein3d.R
\name{translateRangeToColors}
\alias{translateRangeToColors}
\title{translateRangeColors}
\usage{
translateRangeToColors(
  rangeValues,
  insideColor,
  outsideColor = NA,
  inbetween = TRUE
)
}
\arguments{
\item{rangeValues}{numeric vector, range of values: if 'inbetween' is TRUE, then the resulting
function will determine if a value falls in between the first two values. If 'inbetween' is
FALSE, then the resulting function will check if a value is part of the rangeValues argument.
Note that in that case one can work with non-numeric values, see example#'}

\item{insideColor}{color to assign to values that fall within or are part of a range}

\item{outsideColor}{color to assign to values that fall outside or are not part of a range}

\item{inbetween}{logical value that determines the type of range checking}
}
\value{
a function that takes the argument 'value', which it then translates to the correct color
(vectorized)
}
\description{
function factory that generates a color translation function which assigns a color
based on whether the argument value is between two values or if the argument value is part of a
range values
}
\examples{
translateRangeToColors(rangeValues = c(1, 10), insideColor = "white")(1:20)
translateRangeToColors(rangeValues = c(1, 10), insideColor = "white", inbetween = FALSE)(1:20)
translateRangeToColors(rangeValues = 1:10, insideColor = "white", inbetween = FALSE)(1:20)
translateRangeToColors(rangeValues = c(1,2,3, 11, 12, 13), insideColor = "white",
 outsideColor = "red", inbetween = FALSE)(1:20)
translateRangeToColors(rangeValues = c("A","B","C"), insideColor = "white",
 outsideColor = "red", inbetween = FALSE)(LETTERS)
}
