% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{whichModifications}
\alias{whichModifications}
\title{whichModifications}
\usage{
whichModifications(modification, splitCharacter = ",", trim = TRUE)
}
\arguments{
\item{modification}{character vector of format \if{html}{\out{<modification>}} \if{html}{\out{<splitCahracter>}} \if{html}{\out{<modification>}} \if{html}{\out{<splitCahracter>}} etc etc}

\item{splitCharacter}{character vector used to split the modification argument}

\item{trim}{logical vector. Default is TRUE: the result will be trimmed (spaces on both sides removed). When FALSE, this is not done}
}
\value{
character vector
}
\description{
Helper function to list modifications in a character vector of format modification,splitCahracter,modification,splitCahracter, etc etc
}
\examples{
whichModifications('Oxidation,Phospho')
whichModifications('Oxidation , Phospho')
whichModifications('Oxidation , Phospho', trim = FALSE)
whichModifications('Oxidation , Phospho, Carbamidomethyl')
whichModifications(c('Oxidation , Phospho','Carbamidomethyl, Phospho'))
}
